InputGetter + Gpio input support for beaglebone black
InputGetter is a class that updates EffectParameters
instance with new values for button_pressed_times.
button_pressed_times is a list of seconds since last
button press for each button.

One simple implementation of InputGetter is GpioInputGetter.
This class takes a list of GPIOs to monitor as a constructor
parameter, exports GPIOs and checks the state on each
.update() call