ButtonInputGetter
Button input getter checks to see if there's a node
in /dev. If there is - its a sign of a button press.
If node is not present - it means that the button
is not pressed. It is assumed that there's another
daemon process that runs in parallel to control the nodes.