This is a gamma correction table
It is experimental and may or may not produce great results, but we have
high hopes.  It was generated by ../utils/gamma_table_generator.c which
you can modify and compile using gcc -o gamma_table_generator
../utils/gamma_table_generator.c

Additionally, the last element was altered to have the LEDs stay on at
full brightness and therefore were changed to be on = 32, off = 0.