Initial import of upstream library png.py, version 0.0.20
It's getting harder and harder to find a version of png.py that works with
python2.  There used to be an Ubuntu package called "python-png" that provided
it, but it appears to have been dropped from Ubuntu 20.04.  The most recent
version of the upstream library has also dropped support for python2.7.  So,
we'll just embed the last upstream version of png.py that included python2.7
support directly into this repo.  I downloaded the file from
<https://raw.githubusercontent.com/drj11/pypng/pypng-0.0.20/code/png.py>.

Eventually we'll need to convert this codebase from python2.7 to python3, but
for now we can continue limping along...